Performance improvements
- Separate sidebar into 2 smaller components
- Do not show enable dialog for disabling notifications
- MOTD update
- Remove cross origin setting (too many issues for current version)
- update firebase
- update sentry
