+++ package.json
@@ -29,7 +29,7 @@
     "@nuxtjs/recaptcha": "^0.6.2",
     "@nuxtjs/sentry": "^4.1.2",
     "@nuxtjs/toast": "^3.3.0",
-    "@sentry/browser": "^5.16.1",
+    "@sentry/browser": "^5.17.0",
     "@videojs/http-streaming": "^1.13.3",
     "babel-runtime": "^6.26.0",
     "braintree-web-drop-in": "^1.22.1",
@@ -50,8 +50,7 @@
     "videojs-contrib-quality-levels": "^2.0.9",
     "videojs-hls-quality-selector": "^1.1.1",
     "vue-chartjs": "^3.5.0",
-    "vue-clipboard2": "^0.3.1",
-    "vue-virtual-scroller": "^1.0.0-rc.2"
+    "vue-clipboard2": "^0.3.1"
   },
   "devDependencies": {
     "@nuxtjs/style-resources": "^1.0.0",
